Add files via upload
Decorators in Python are a powerful and flexible way to modify or extend the behavior of functions or methods. A decorator is a function that takes another function as an argument and typically returns a new function that usually extends or modifies the behavior of the original function. Decorators are often used for code reuse, logging, access control, and more.

I added Python file for decorator with simple example